move validate lend/borrow functions to validate.go
